requirements.txt
Django==2.0.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
postgres==2.2.1
psycopg2==2.7.4
PyJWT==1.6.0
pytz==2018.3
